After eliminate 3inarow, can't replace with the same 3
The problem this was trying to solve was this:

After eliminating a three in a row, it sometimes happens that the three randomly generated replacement dots are the exact same dots that were eliminated. This is confusing.

I wanted to specfiy that the replacement dots cant' be the exact same three dots, but GPT4 generated something which was slightly different, but still did the job. So this change means that NONE of the three replacement dots can be the same as the three-in-a-row colour.

I also tidied up (ie deleted out) some old functions which have been commented out for a while