Slide moves needed displayed on front page
The overall change here is about game progression: i.e. the number of times you need to eliminate a 3-in-a-row increases over the course of the game.

I included some info on the front page. Originally for debugging pruposes, but I left it here since I think it might make things clearer for players.