Game tells you when it's game over
Doing this ended up taking ages. It involved creating the hasPOssibleMoves function, but this then caused lots fo bugs. I can't remember excctly how they were solved but I think ti was something to do with tweaking the generateGrid function to ensure that objects had the right properties (err, I think it was ensuring that the colour property was attached to dot objects). There was also something about the updateDotDataAttributes function, which had a bug in it which is now sorted.

Because of all the faffing, I can't remember exactly which changes were made to make this happen.